docker/provisioner/topics/requirements.txt
requests
kafka-python
cfnresponse